% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enum}
\alias{enum}
\title{Enumerate elements inline, in bullet-list, or in numbered-list form.}
\usage{
enum(xs, sep = ",", bullet = NULL, init = NULL)
}
\arguments{
\item{xs}{a vector of elements to be enumerated}

\item{sep}{a character string to separate the elements if inline}

\item{bullet}{possibly bool or a character string. If \code{TRUE}, returns a
bullet list with default bullet points. If a character string is
given, uses the string as bullet points.}

\item{init}{possibly bool or a number. If \code{TRUE}, returns a numbered list
with default numbers. If a number is given, uses the number to start
enumeration with.}
}
\value{
a character string enumerating the elements
}
\description{
This is a wrapper for functions \code{enum_inline}, \code{enum_bullet}, and
\code{enum_numbered}. If \code{bullet} or \code{init} option is neither \code{NULL} nor
\code{FALSE}, then it generates a bullet list or a numbered list appropriately.
Otherwise, it enumerates the elements inline.
If both \code{bullet} and \code{init} options are neither \code{NULL} nor \code{FALSE}, a
numbered list precedes a bullet list.
}
\examples{
enum(1:3) #=> "1, 2, and 3"
enum(1:3, bullet = TRUE) #=> "- 1\\n- 2\\n- 3"
enum(1:3, init = 2) #=> "2. 1\\n3. 2\\n4. 3"
}
