% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enum_inline}
\alias{enum_inline}
\title{Enumerate elements inline.}
\usage{
enum_inline(xs, sep = ",")
}
\arguments{
\item{xs}{a vector of elements to be enumerated}

\item{sep}{a character string to separate the elements}
}
\value{
a character string enumerating the elements
}
\description{
Enumerate elements inline.
}
\examples{
enum_inline(1:2) #=> "1 and 2"
enum_inline(1:3) #=> "1, 2, and 3"
enum_inline(1:3, sep = ";") #=> "1; 2; and 3"
}
