% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enum_numbered}
\alias{enum_numbered}
\title{Enumerate elements in a numbered list.}
\usage{
enum_numbered(xs, init = 1)
}
\arguments{
\item{xs}{a vector of elements to be enumerated}

\item{init}{initial number to begin enumeration}
}
\value{
a character string enumerating the elements in a numbered list
}
\description{
Enumerate elements in a numbered list.
}
\examples{
enum_numbered(1:3) #=> "1. 1\\n2. 2\\n3. 3"
enum_numbered(1:3, init = 10) #=> "10. 1\\n11. 2\\n12. 3"
}
