% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_prefix}
\alias{guess_prefix}
\title{Common prefix of a given vector of strings without decimals in its tail.}
\usage{
guess_prefix(ss)
}
\arguments{
\item{ss}{a vector of strings}
}
\value{
the common prefix without decimals in its tail (string)
}
\description{
Common prefix of a given vector of strings without decimals in its tail.
}
\examples{
guess_prefix(c("a11", "a12", "a13")) #=> "a"
}
